Modified uploader to create tables in advance to set tableIds.
Tablets are created using an assumed table ID (because tableIDs are used in the
hash), and so when uploading tablets, it's important that tables have the
expected table IDs...

Now the uploader will ceate the tables in advance in the correct order so as to
get the correct tableIDs. This is, of course, totally janky because it's
dependent on RAMCloud table creation giving us the table IDs that we expect, but
it's the best we can do at the moment since RAMCloud doesn't offer us the option
of requesting specific tableIDs.